SQUASH: pinctrl: fix jz4730 lcd table
modes must be an u8 array and not an int array

+++ drivers/pinctrl/pinctrl-ingenic.c
@@ -365,7 +365,7 @@ static int jz4730_nand_cs5_pins[] = { 0x57, };
 static int jz4730_pwm_pwm0_pins[] = { 0x5e, };
 static int jz4730_pwm_pwm1_pins[] = { 0x5f, };
 
-static int jz4730_lcd_8bit_funcs[] = { 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, };
+static u8 jz4730_lcd_8bit_funcs[] = { 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, };
 
 static const struct group_desc jz4730_groups[] = {
 	INGENIC_PIN_GROUP("mmc-1bit", jz4730_mmc_1bit, 1),
